Update dbn_impl.hpp
Removed commented code. The methods use a custom generator type reg_generator_t which currently matches categorical_generator_t.

Only issue is the return value.

return trainer.train(*this, generator, max_epochs);

For regression, average loss seems like a good fit, but the trainer is treating it like a categorical generator, with classification error and nan loss. So the return value and the printed output is invalid, but as far as I can tell the training does work.